% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterCrossV_functions.R
\name{AIC_indices}
\alias{AIC_indices}
\title{Function that returns the AIC of a list of models
In the case of a tweedie model (TweedGLM), it also returns the XI value to be used in cross-validation if not fixed previously}
\usage{
AIC_indices(
  x,
  Y_data_sample,
  Models_tmp_nb,
  modeltype,
  fixXI,
  Y_data_sample_lcc = NA,
  MaxDist = NA,
  Phi = NA,
  Model,
  fix.lambda,
  lambda
)
}
\arguments{
\item{x}{the model number to be fitted. From 1 to length(Models_tmp_nb).}

\item{Y_data_sample}{data.frame or SpatialPointsDataFrame of observations with covariates}

\item{Models_tmp_nb}{matrix with column of model formulas as character}

\item{modeltype}{sub-model type}

\item{fixXI}{Power of the Tweedie model (\code{\link[statmod]{tweedie}})}

\item{Y_data_sample_lcc}{dataset to be fit on, with projected CRS
("Krige*" modeltype only)}

\item{MaxDist}{Maximum distance for variogram ("Krige*" modeltype only)}

\item{Phi}{Range for Phi fitting ("Krige*" modeltype only)}

\item{Model}{Model type for variogram ("Krige*" modeltype only)}

\item{fix.lambda}{logical, indicating whether the Box-Cox transformation parameter
lambda should be regarded as fixed (fix.lambda = TRUE) or should be be estimated
(fix.lambda = FALSE). Defaults to TRUE.}

\item{lambda}{value of the Box-Cox transformation parameter lambda.
Regarded as a fixed value if fix.lambda = TRUE otherwise as the initial value
for the minimisation algorithm. Defaults to 1. Two particular cases are
lambda = 1 indicating no transformation and lambda = 0 indicating log-transformation.}
}
\description{
Function that returns the AIC of a list of models
In the case of a tweedie model (TweedGLM), it also returns the XI value to be used in cross-validation if not fixed previously
}
