% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{CovarExtract}
\alias{CovarExtract}
\title{Function to extract covariates from different rasters}
\usage{
CovarExtract(x, cov.paths, names.cov, is.factor)
}
\arguments{
\item{x}{SpatialPointsDataFrame}

\item{cov.paths}{vector of paths of all raster files to be added in the multi-layer raster}

\item{names.cov}{Optional. vector of names of multi-layer raster. Same length as x.}

\item{is.factor}{position of covariates that are factors or categories.
Resampling (if required) is bilinear by default but "ngb" (nearest neighbour) for factors.}
}
\description{
Original covariates raster may have different projections.
CovarExtract allows to extract in these different projections.
}
