% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{CreateExtent}
\alias{CreateExtent}
\title{Find area extent
Verify if this function is not in my spatial script}
\usage{
CreateExtent(x, saveWD)
}
\arguments{
\item{x}{SpatialPointsDataFrame}

\item{saveWD}{directory where to save Extent}
}
\description{
Find area extent
Verify if this function is not in my spatial script
}
