% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{Fig_split}
\alias{Fig_split}
\title{Find an equilibrated layout for a figure according to number of panels to include}
\usage{
Fig_split(x)
}
\arguments{
\item{x}{number of panels of the future plot}
}
\value{
nrow
}
\description{
Find an equilibrated layout for a figure according to number of panels to include
}
\details{
As screens are larger than height, nrow is set not to exceed ncol.
Rules are : ncol >= nrow, ncol <= nrow + 2, ntot >= x
100*100 is the maximum window. It is surely too big for a screen, but who
knows what you do with this function... Be sure your window is big enough !
}
