% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterCrossV_functions.R
\name{fit_model}
\alias{fit_model}
\title{Function to fit a model among all possible}
\usage{
fit_model(x, model, fixXI, modeltype, x_lcc, MaxDist, Phi, Model,
  fix.lambda, lambda)
}
\arguments{
\item{x}{data.frame or SpatialPointsDataFrame of observations with covariates}

\item{model}{model formula written as character}

\item{fixXI}{Power of the Tweedie model (\code{\link[statmod]{tweedie}})}

\item{modeltype}{sub-model type}

\item{x_lcc}{Projected version of x for kriging}

\item{MaxDist}{Maximum distance for variogram ("Krige*" modeltype only)}

\item{Phi}{Range for Phi fitting ("Krige*" modeltype only)}

\item{Model}{Model type for variogram ("Krige*" modeltype only)}

\item{fix.lambda}{logical, indicating whether the Box-Cox transformation parameter
lambda should be regarded as fixed (fix.lambda = TRUE) or should be be estimated
(fix.lambda = FALSE). Defaults to TRUE.}

\item{lambda}{value of the Box-Cox transformation parameter lambda.
Regarded as a fixed value if fix.lambda = TRUE otherwise as the initial value
for the minimisation algorithm. Defaults to 1. Two particular cases are
lambda = 1 indicating no transformation and lambda = 0 indicating log-transformation.}
}
\description{
Function to fit a model among all possible
}
