% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{get_opt_value}
\alias{get_opt_value}
\title{Internal function for global options}
\usage{
get_opt_value(name)
}
\description{
Internal function for global options
}
\seealso{
\code{\link{modelselect_opt}}
}
\keyword{internal}
