% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{is_ProjEqual}
\alias{is_ProjEqual}
\title{Test if projections are equals}
\usage{
is_ProjEqual(x, y)
}
\arguments{
\item{x}{proj4string}

\item{y}{proj4string}
}
\description{
Some projection are the same but written differently
or with less useless arguments.
This function test if two projections are equal while accounting for
multiple writings, in particular when EPSG can not be retrieved
}
