% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\docType{data}
\name{lt}
\alias{lt}
\title{Internal function for global options}
\format{An object of class \code{list} of length 2.}
\usage{
lt
}
\description{
Internal function for global options
}
\seealso{
\code{\link{modelselect_opt}}
}
\keyword{internal}
