% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{model_select}
\alias{model_select}
\title{Characteristics of the model on which to produce outputs}
\usage{
model_select(saveWD, new.data, Num = NULL, model = NULL, modeltype = NULL,
  powerXI = NULL)
}
\arguments{
\item{saveWD}{directory where outputs of model selection are saved, including
AllModels_crossV_ordered.}

\item{new.data}{SpatialPointsDataFrame dataset used to fit the model.}

\item{Num}{numeric The 'Num' value of the model retained for further analysis
like those saved in AllModels_crossV_ordered. Not useful if model and Family
are set. Default to the first model in AllModels_crossV_ordered.}

\item{model}{character string A user-defined model specification. This requires
to set the Family parameter. This overrides effects of 'Num', however, if the
model exists with the exact same expression in the list of model fitted,
'Num' will be re-attributed.}

\item{modeltype}{The model type as chosen among modeltypes available
(see \code{\link{modelselect_opt}})}

\item{powerXI}{XI parameter for a tweedie model.}
}
\value{
Num, modeltype, Family, formulaX
}
\description{
Characteristics of the model on which to produce outputs
}
