% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{ModelOrder}
\alias{ModelOrder}
\title{Order fitted models}
\usage{
ModelOrder(saveWD, plot, zip.file = TRUE, cl = NULL)
}
\arguments{
\item{saveWD}{directory where outputs of the cross-validation procedure have
been saved or zip.file of this directory.
All necessary information have been saved in this directory to be
able to compile results. This is the only output of findBestModel in the
global R environment.}

\item{plot}{logical. Either to create the output plots or not. Saved
directly in saveWD.}

\item{zip.file}{TRUE (default) to save all outputs in a zipfile with saveWD,
FALSE for no zipfile output, or path to a new zip file}

\item{cl}{a cluster as made with \code{\link[snow]{makeCluster}}.
If empty, nbclust in \code{\link{modelselect_opt}} will be used.}
}
\value{
VeryBestModels_crossV: Best models among all
BestForModeltypes:  Best models among all in the modeltypes type
Figures are produced to visualize ranking and eventually help for a choice
}
\description{
Find the models that outperform the others in terms of cross-validation index
}
\details{
According to the family of model, the index of cross-validation procedure is
selected to rank all models fitted for all modeltypes conjointly. Model ranks
are then statistically compared. Model with lowest ranks not statistically
different than the first ranked model are kept.
}
