% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results.R
\name{ModelResults}
\alias{ModelResults}
\title{Study the outputs and predictions of a specific model}
\usage{
ModelResults(saveWD, plot, Num = NULL, model = NULL, modeltype = NULL,
  powerXI = NULL, Marginals = FALSE, zip.file = TRUE, cl = NULL,
  n.ech.marginal = 5000)
}
\arguments{
\item{saveWD}{directory where outputs of the cross-validation procedure have
been saved or zip.file of this directory.
All necessary information have been saved in this directory to be
able to compile results. This is the only output of findBestModel in the
global R environment.}

\item{plot}{logical Whether to do plot outputs or not. Although there are not
a lot of other types of outputs...}

\item{Num}{numeric The 'Num' value of the model retained for further analysis
like those saved in AllModels_crossV_ordered. Not useful if model and Family
are set. Default to the first model in AllModels_crossV_ordered.}

\item{model}{character string A user-defined model specification. This requires
to set the Family parameter. This overrides effects of 'Num', however, if the
model exists with the exact same expression in the list of model fitted,
'Num' will be re-attributed.}

\item{modeltype}{The type of the model as chosen among modeltypes available
(see \code{\link{modelselect_opt}})}

\item{powerXI}{XI parameter for a tweedie model.}

\item{Marginals}{Calculate full marginal prediction for each covariate}

\item{zip.file}{TRUE (default) to save all outputs in a zipfile with saveWD,
FALSE for no zipfile output, or path to a new zip file}

\item{cl}{a cluster as made with \code{\link[snow]{makeCluster}}.
If cl is empty, nbclust in \code{\link{modelselect_opt}} will be used
to create a cluster.}

\item{n.ech.marginal}{5000 by default. Number of random samples for marginals
estimations.}
}
\description{
Study the outputs and predictions of a specific model
}
