% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{prec_data}
\alias{prec_data}
\title{Find the most precise data of a vector}
\usage{
prec_data(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
precision as a power of 10: 10^n
}
\description{
Find the most precise data of a vector
}
