% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CartoPred.R
\name{predR}
\alias{predR}
\title{Predictions in response and link scales}
\usage{
predR(object, newdata)
}
\arguments{
\item{object}{fitted model as output of \code{\link{glm}} or \code{\link{gam}}}

\item{newdata}{a data frame in which to look for variables with which to predict}
}
\description{
Internal function to be coupled with \code{\link[raster]{predict}}
and \code{\link[raster]{clusterR}}
}
