% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{Prepare_covarStack}
\alias{Prepare_covarStack}
\title{Merge covariate rasters from different extents and projections into a multi-layer Raster}
\usage{
Prepare_covarStack(
  cov.paths,
  r.ref,
  ext,
  outWD,
  names.cov,
  is.factor,
  verbose = 1,
  cl = NULL
)
}
\arguments{
\item{cov.paths}{vector of paths of all raster files to be added in the multi-layer raster}

\item{r.ref}{Raster used as a reference for extent, resolution, crs.
Default is the first one of the list.}

\item{ext}{Optional object of class extent to crop r.ref before stack.}

\item{outWD}{Directory where to save new reprojected, cropped rasters.
Default to "tempdir()/extent".}

\item{names.cov}{Optional. vector of names of multi-layer raster. Same length as x.}

\item{is.factor}{position of covariates that are factors or categories.
Resampling (if required) is bilinear by default but "ngb" (nearest neighbour) for factors.}

\item{verbose}{level of messages shown}

\item{cl}{a cluster as made with \code{\link[parallel]{makeCluster}}.
If cl is empty, nbclust in \code{\link{modelselect_opt}} will be used
to create a cluster.}
}
\description{
Merge covariate rasters from different extents and projections into a multi-layer Raster
}
\details{
Rasters are merged in a unique RasterStack. If the projection of
a raster is different than the reference (r.ref), it is reprojected. If the
extent is different, it is resampled. If the resolution is smaller than
0.9*res, it is first aggregated, then resampled.
Method depends if projection is really different,
otherwise this is only a change in name of proj
}
