% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{RefRasterize}
\alias{RefRasterize}
\title{Create a RefRaster for regular grid dataset resampling}
\usage{
RefRasterize(x, res, degree = FALSE)
}
\arguments{
\item{x}{SpatialPointsDataFrame}

\item{res}{raster resolution targeted. Vector of one or two values for x and y resolutions.}

\item{degree}{Unit of the resolution. TRUE for degrees, FALSE for meters. Default to FALSE
Size one or two, if different resolution for x and y.}
}
\description{
Create a RefRaster for regular grid dataset resampling
}
