% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolFunctions.R
\name{Rm_WhiteMargins}
\alias{Rm_WhiteMargins}
\title{A function to crop white margins of a PNG image}
\usage{
Rm_WhiteMargins(x, pixel = 2)
}
\arguments{
\item{x}{path to the PNG image}

\item{pixel}{number of white pixels lines to keep}
}
\description{
A function to crop white margins of a PNG image
}
