% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreparation.R
\name{spatialcor_dist}
\alias{spatialcor_dist}
\title{Figure to see spatial autocorrelation that may allow to define the grid
size for gridded procedure}
\usage{
spatialcor_dist(
  x,
  y,
  longlat = FALSE,
  max1,
  lag1,
  max2,
  lag2,
  binomial = TRUE,
  thd = TRUE,
  plot = TRUE,
  saveWD,
  figname,
  simplify.grid = FALSE
)
}
\arguments{
\item{x}{A SpatialPointsDataFrame}

\item{y}{The column number on which to calculate correlation.
Otherwise a column should be names "dataY"}

\item{longlat}{logical. longlat data or not. Default to FALSE}

\item{max1}{Numeric. maximum distance of panel 1 in meters.
Default to max distance divided by 3.}

\item{lag1}{Numeric. step distance for calculating autocorrelation in meters.
Default to max1/100.}

\item{max2}{Numeric. maximum distance of panel 1 in meters. Default to max1/10.}

\item{lag2}{Numeric. step distance for calculating autocorrelation in meters.
Default to max2/100.}

\item{binomial}{Logical. Presence-absence data (TRUE) or continuous data (FALSE).}

\item{thd}{logical. If TRUE, the function suggest a distance threshold according to
breakpoint in correlation. This is only a suggestion. nls function is used.}

\item{plot}{Logical.}

\item{saveWD}{directory where to save the output figure. If null, figure appears on screen.}

\item{figname}{character. If saveWD is not empty, you can specify the name of
the output figure (without extension). Default to "Correlogram".}

\item{simplify.grid}{logical. If dataset is too big, it will be difficult to calculate
all distances between all points. A simplification is to divide the area into a grid
and calculate distances and correlation in each cell of the grid separately.
Results are then merged. Grid used depends on the largest of max1 and max2 values.
If less than 10 values, there are randomly merged with another cell}
}
\value{
A figure of correlation is showed or saved (in saveWD).
A threshold is suggested if thd=TRUE
}
\description{
Figure to see spatial autocorrelation that may allow to define the grid
size for gridded procedure. Two figures are produced with different distance step.
}
