% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot_data.R
\docType{data}
\name{gplot_data}
\alias{gplot_data}
\title{Transform raster as data.frame to be later used with ggplot
Modified from rasterVis::gplot}
\format{An object of class \code{function} of length 1.}
\usage{
gplot_data
}
\description{
See \code{cartomisc::\link[cartomisc]{gplot_data}} for details.
}
\keyword{datasets}
