% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot_data.R
\name{gplot_data}
\alias{gplot_data}
\title{Transform raster as data.frame to be later used with ggplot
Modified from rasterVis::gplot}
\description{
See \code{cartomisc::\link[cartomisc]{gplot_data}} for details.
}
