% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_config.R
\name{lt}
\alias{lt}
\title{Internal function for global options}
\usage{
lt(..., RESET = FALSE, READ.ONLY = NULL, LOCAL = FALSE,
  ADD = FALSE)
}
\description{
Internal function for global options
}
\seealso{
\code{\link{modelselect_opt}}
}
\keyword{internal}
